Disable reordering task groups for now
Turns out, WinUI 3 is not helping me here. List reorders aren't
presented via the `ObservableCollection` by firing a `Move` action in
`NotifyCollectionChangeArgs`, but instead a removal and re-insertion is
fired. Which for some is probably fine, but not for me, as I wanted to
use the action as the source of truth wrt what was actually done to the
collection, but given this behaviour, I can no longer distinguish a
user-requested removal from the list from a standard reorder.

I can see a few ways to address this but they are all a bit of a waste
of time given where I want to get to with the MVP of this app. So
disabling for now. I can deal with the faff later.